+++ tests/inputs/semicolon-1.lua
@@ -1,1 +1,2 @@
 local x = 1; -- comment
+return x; -- comment