+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed trailing comments at the end of multiline tables being lost
 - Fixed panic "stmt trailing comments not implemented" occuring due to incomplete function
+- Fixed trailing comments after semicolons at the end of last statements being lost when formatting
 
 ## [0.6.0] - 2021-03-27
 ### Added