+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Further improvements to the way binary expressions are hung on new lines
+
 ### Fixed
 - Fixed trailing comments at the end of multiline tables being lost
 - Fixed panic "stmt trailing comments not implemented" occuring due to incomplete function