+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added hanging for chained function calls. See [#109](https://github.com/JohnnyMorganz/StyLua/issues/109)
+- Long function definitions (normally with parameters containing types and a return type) will now be split across multiple lines if they surpass the column limit
 
 ### Changed
 - Further improvements to the way binary expressions are hung on new lines
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed comments at the end of tables being indented one extra level
 - Fixed trailing comments within if-elseif-else blocks not being correctly indented.
 - Fixed `do` in a `while ... do` statement not correctly indented when the condition spans multiple lines
+- Fixed multiline parameters for a function definition inside of an indent block (e.g. a table) not being correctly indented
 
 ## [0.6.0] - 2021-03-27
 ### Added