+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.0] - 2021-04-13
 ### Added
 - Added hanging for chained function calls. See [#109](https://github.com/JohnnyMorganz/StyLua/issues/109)
 - Long function definitions (normally with parameters containing types and a return type) will now be split across multiple lines if they surpass the column limit